mysite/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.0
billiard==3.3.0.23
boto3==1.16.22
botocore==1.19.22
celery==3.1.18
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.3
django-celery-results==2.0.0
django-decouple==2.1
Django-Gtts==0.3
django-storages==1.10.1
dnspython==1.16.0
gTTS==2.2.1
idna==2.10
jmespath==0.10.0
kombu==3.0.37
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.4
pywin32==300
redis==3.2.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
